pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pytket-custatevec"
version = "0.1.2"
description = "Extension for pytket providing access to cuStateVec functionality."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Fabian Finger", email = "fabian.finger@quantinuum.com"},
  { name = "Gabriel Matos", email = "gabriel.matos@quantinuum.com"},
  { name = "Pranav Kalidindi", email = "pranav.kalidindi@quantinuum.com"},
]

dependencies = [
  "pytket",
  "cupy-cuda12x",
  "cuquantum-python-cu12>=25.3.0",
  "plotly>=6.5.0",
]

[project.optional-dependencies]
# Core development tools (Linting, Formatting, Typing)
dev = [
  "pre-commit>=4.4.0",
  "pytest>=8.4.0",
  "ruff>=0.14.0",
  "mypy>=1.18.0",
]

# Documentation tools (Matches your CI workflow)
docs = [
  "mkdocs",
  "mkdocs-material",
  "mkdocstrings[python]",
  "mike",  # For versioning
]

# Dependencies required to run the full test suite
test = [
  "pytest",
  "pytket-qiskit",
  "pytket-qulacs",
]

[tool.setuptools.packages.find]
where = ["."]

[project.urls]
Repository = "https://github.com/CQCL/pytket-custatevec.git"

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.refurb]
python_version = "3.10"
ignore = ["FURB184"]

[tool.typos.default.extend-words]
aer = "aer"
